Add 3 queues (free op, busy op and calls).
Every call is added to the queue, and only removed if there is a
operator ready to answer it. If there is not enough operators to answer
the calls.

When the operator answer the call, it moves from the free queue to the
busy queue. The vice-versa operation, when the caller disconnects, and
the operator must be returned to the free queue, will be added later.