Use DeferredQueue instead of a custom solution.
DeferredQueue is a queue which gets a Deferred when there are no items
left in it. The previous implementation was non-functional, had a
serious race condition and didn't scale (since every call waiting for an
operator would have a thread). It also simplifies implementation (less
code).

The Deferreds calls its callbacks in the order in which they are
declared, e.g., if call 1 and call 2 have a Deferred for an item in the
queue, call 1 will be served first, and then call 2. It's a natural
queue, and there is no racing, since Deferreds fires only once.